+++ workbench/device/default_devices/cactus_typical_devices.csv
@@ -1,0 +1,10 @@
+,bishop_N_p_typical,bishop_N_s_typical,bishop_breakdown_exp,bishop_breakdown_factor,bishop_breakdown_voltage,desoto_bandgap_temp_coe_dEgdT,desoto_diode_factor_a_ref,desoto_energy_bandgap_Egref,desoto_photocurrent_I_L_ref,desoto_saturation_current_I_o_ref,desoto_series_resist_R_s_ref,desoto_short_circuit_temp_coe_alpha_sc,desoto_shunt_resist_R_sh_ref,general_cell_area_mm2_typical,general_cell_height_mm_typical,general_cell_type,general_cell_width_mm_typical,general_description,general_heterojunction,general_monolithic,module_n_subcell_cols_typical,module_n_subcell_rows_typical,module_shape_N_columns_typical,module_shape_N_rows_typical,module_shape_area_m2,module_shape_column_spacing_mm,module_shape_flex,module_shape_frame_thickness_mm,module_shape_height_mm,module_shape_row_spacing_mm,module_shape_width_mm,performance_NOCT_T_degC,performance_cell_efficiency_ref,performance_module_efficiency_ref,performance_power_W_ref,performance_temp_coe_gamma_pctC,performance_thermal_resilience,general_device_id
+monoSi_singleJ_halfcut,2.0,72.0,3.0,0.002,-15.0,-0.0002677,1.82855741110967,1.121,11.342993908696188,1.7707709852891217e-11,0.2026238233810742,0.00034,767.4762927799923,13778.0,83.0,monoSi,166.0,typical half cut monocrystalline module,False,False,1.0,1.0,6.0,24.0,2.19232,4.0,False,4.0,2108.0,4.0,1040.0,45.6,0.227,0.206,451.0219251206661,-0.347,0.7378777346506352,D001
+monoSi_singleJ_square,1.0,60.0,3.0,0.002,-15.0,-0.0002677,1.529236889928173,1.121,10.569788965147488,5.899046896424323e-11,0.2138122487826511,0.00042,230.65332842402825,26244.0,162.0,monoSi,162.0,typical square cell monocrystalline module,False,False,1.0,1.0,6.0,10.0,1.764004,6.0,False,7.0,1706.0,6.0,1034.0,45.1,0.207,0.185,325.70374801908235,-0.381,0.7345650601516399,D002
+polySi_singleJ_halfcut,2.0,72.0,3.0,0.002,-15.0,-0.0002677,1.7588625313277315,1.121,9.826715600700185,3.031863068048801e-11,0.3269776426574218,0.0005,478.1285649339557,13778.0,83.0,polySi,166.0,typical polycrystalline halfcut cell module,False,False,1.0,1.0,6.0,24.0,2.19232,4.0,False,4.0,2108.0,4.0,1040.0,45.4,0.179,0.162,354.4958848831378,-0.362,0.7380985227710395,D003
+polySi_singleJ_square,1.0,72.0,3.0,0.002,-15.0,-0.0002677,1.851403851078951,1.121,9.398649540129544,1.2758406719334542e-10,0.309676937271745,0.00049,336.1874227901243,25201.5625,158.75,polySi,158.75,typical square cell polycrystalline module,False,False,1.0,1.0,6.0,12.0,1.9149245,2.0,False,5.0,1957.0,3.0,978.5,45.2,0.185,0.175,335.15983337309206,-0.409,0.7298069010438926,D004
+cigs_singleJ_wafer,1.0,60.0,3.0,0.002,-15.0,-0.0002677,1.341637263719694,1.15,7.451246955616755,2.1324413857187347e-11,0.6340979581320793,-0.00012,26.956585221740752,26569.0,163.0,cigs,163.0,typical CIGS module,False,False,1.0,1.0,6.0,10.0,1.66,2.0,True,3.0,1660.0,2.0,1000.0,55.9,0.104,0.099,164.99812881412214,-0.385,0.7154025021437951,D005
+cdte_singleJ_monolith,1.0,264.0,3.0,0.002,-15.0,-0.0003,7.987883856729444,1.475,2.545000718407697,2.969474195982689e-12,4.901760348659221,0.0005600000000000001,2489.736533489824,9217.5,7.5,cdte,1229.0,typical CdTe module,False,True,5.0,1.0,1.0,264.0,2.4484761,0.1,False,1.5,1984.5,0.0,1233.8,50.7,0.177,0.176,430.9359901617367,-0.307,0.6851101365597041,D006
+monoSi_tandemJ_halfcut,2.0,60.0,3.0,0.002,-15.0,-0.0002677,1.518287640982106,1.121,10.24232266224336,2.4054240043665842e-12,0.1554028954496397,0.00037,129.01202594262088,13778.0,83.0,monoSi,166.0,typical HJT halfcut monocrystalline module,True,False,1.0,1.0,6.0,20.0,1.751661,1.0,False,5.0,1719.0,2.0,1019.0,44.0,0.218,0.206,360.03453325854275,-0.279,0.7509200872800227,D007
+monoSi_tandemJ_square,1.0,96.0,3.0,0.002,-15.0,-0.0002677,2.339992667765493,1.121,6.023251974583235,4.580844910084929e-13,0.7246887020833652,0.00037999999999999997,328.6588781886511,16900.0,130.0,monoSi,130.0,typical HJT square monocrystalline module,True,False,1.0,1.0,8.0,12.0,1.6776678,1.3,False,3.0,1584.2,1.3,1059.0,44.9,0.204,0.197,330.3998503936697,-0.257,0.7569566109666167,D008
+monoSi_singleJ_square_mini,1.0,6.0,3.0,0.002,-15.0,-0.0002677,2.339992667,1.121,8.628778,4.956246e-10,0.300444,3.784e-05,89.785065,22500.0,150.0,monoSi,150.0,a small monocrystalline module that is used for the template file,False,False,1.0,1.0,2.0,3.0,0.16745,2.5,False,15.0,492.5,2.5,340.0,45.4,2.603,2.099,351.47183162090073,-0.00485,0.0,D009
